+++ packages/evmos-wallet/src/internal/wallet/functionality/networkConfig.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
 export const EVMOS_RPC_URL =
-  process.env.NEXT_PUBLIC_EVMOS_RPC_URL ?? "https://eth.bd.evmos.org:8545/";
+  process.env.NEXT_PUBLIC_EVMOS_RPC_URL ?? "https://evmos.lava.build/";
 export const EVMOS_GRPC_URL =
-  process.env.NEXT_PUBLIC_EVMOS_GRPC_URL ?? "https://rest.bd.evmos.org:1317/";
+  process.env.NEXT_PUBLIC_EVMOS_GRPC_URL ?? "https://rest.evmos.lava.build/";
 export const EVMOS_CHAIN_NAME =
   process.env.NEXT_PUBLIC_EVMOS_CHAIN_NAME ?? "Evmos";
 export const EVMOS_SYMBOL = process.env.NEXT_PUBLIC_EVMOS_SYMBOL ?? "EVMOS";