+++ packages/ui-helpers/src/PrimaryButton.tsx
@@ -28,7 +28,7 @@ export const PrimaryButton = ({
         {
           "bg-red text-pearl hover:bg-red1 active:bg-red2":
             variant === "primary",
-          "border-2 border-pink-300 gradient w-full text-base md:text-lg rounded-md capitalize font-medium":
+          "border-2 border-pink-300 gradient w-full text-base md:text-lg rounded-md font-medium":
             variant === "outline-primary",
           "bg-red text-pearl hover:bg-red1 active:bg-red2 w-full text-base md:text-lg capitalize font-medium":
             variant === "primary-lg",