+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (fix) fse-754 | packages/ui-helpers 1.0.16 | Add attribution to Coingecko
 - (fix) | apps/assets 1.0.30 packages/icons 1.0.7 packages/ui-helpers 1.0.15 | Transaction failure message even if successful
 
 ## 1.3.3 - 2023-10-03